Accomodate providers with default deletion policies and provider configs
https://github.com/crossplane/crossplane-runtime/pull/255

The defaulting of these fields will be introduced in crossplane-runtime v0.14
per the above PR. We'd like to support a transition period during which a
provider may either specify the default or not. Unfortunately it's not easy to
use cmp to test for "must be unset or value X" so I've just ignored the field
for the time being.

Signed-off-by: Nic Cope <negz@rk0n.org>